2 cambios inscripciondata y carga notas
Inscripcion data tenia mal un dato seteaba en int en vez de double por eso redondeaba a entero los resultados. Cargar notas, agregue un metodo en el constructor para que las tablas no sean editables

+++ src/misionImposible/AccesoADatos/InscripcionData.java
@@ -102,7 +102,7 @@ public class InscripcionData {
                 Inscripcion inscripcion = new Inscripcion();
 
                 inscripcion.setIdInscripto(rs.getInt("idInscripto"));
-                inscripcion.setNota(rs.getInt("nota"));
+                inscripcion.setNota(rs.getDouble("nota"));
 
                 Alumno a = aluData.buscarAlumno(rs.getInt("idAlumno"));
                 inscripcion.setAlumno(a);//COLOCA LA INF DEL ALUMNO PASADO POR ID
